// src/types/generated/BroadcastResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AppError } from "./AppError";
import type { ExecOutput } from "./ExecOutput";

export type BroadcastResult = { broadcastId: string, serverId: string, serverName: string, output: ExecOutput | null, error: AppError | null, };